requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# HTTP requests and async
aiohttp>=3.8.0
requests>=2.31.0
asyncio-throttle>=1.0.0

# GTFS and protobuf
gtfs-realtime-bindings>=1.0.0
protobuf>=4.23.0

# Kafka
confluent-kafka>=2.0.0
kafka-python>=2.0.0

# BigQuery and data warehouse
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
db-dtypes>=1.1.0

# Configuration and environment
python-dotenv>=1.0.0
pyyaml>=6.0.0
pydantic-settings>=2.0.0

# Logging and monitoring
structlog>=23.1.0
prometheus-client>=0.17.0

# Data validation and schemas
pydantic>=2.0.0
marshmallow>=3.20.0

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0

# CLI interface
click>=8.1.0

# Dashboard and web framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development tools
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Orchestration (choose one)
# apache-airflow>=2.7.0
# prefect>=2.10.0

# Optional: Map visualization
folium>=0.14.0
geopandas>=0.13.0
